extracted EventList component and logic from app.vue -Created a new EventList component to display events. -Moved the event fetching and loading logic from app.vue to EventList.vue. -Refactored the event rendering with EventCard and LoadingEventCard based on loading state. -Emitted register event when a user registers for an event
