+++ src/App.vue
@@ -2,22 +2,7 @@
   <main class="container mx-auto my-8 space-y-8">
     <h1 class="text-4xl font-medium">Event Booking Manager</h1>
     <h2 class="text-2xl font-medium">All Events</h2>
-    <section class="grid grid-cols-2 gap-8">
-      <template v-if="!eventsLoading">
-        <EventCard
-          v-for="event in events"
-          :key="event.id"
-          :title="event.title"
-          :when="event.date"
-          :description="event.description"
-          @register="handleRegistration(event)"
-        />
-      </template>
-
-      <template v-else>
-        <LoadingEventCard v-for="i in 4" :key="i" />
-      </template>
-    </section>
+    <EventList @register="handleRegistration($event)" />
 
     <h2 class="text-2xl font-medium">Your Bookings</h2>
     <section class="grid grid-cols-1 gap-4">
@@ -41,26 +26,12 @@
 <script setup>
 import { ref, onMounted } from 'vue';
 
-import EventCard from '@/components/EventCard.vue';
 import BookingItem from '@/components/BookingItem.vue';
-import LoadingEventCard from '@/components/LoadingEventCard.vue';
 import LoadingBookingItem from '@/components/LoadingBookingItem.vue';
+import EventList from '@/components/EventList.vue';
 
-const events = ref([]);
-const eventsLoading = ref(false);
 const bookings = ref([]);
 const bookingsLoading = ref(false);
-
-// Fetch events from mock API
-const fetchEvents = async () => {
-  eventsLoading.value = true;
-  try {
-    const response = await fetch('http://localhost:3001/events');
-    events.value = await response.json();
-  } finally {
-    eventsLoading.value = false;
-  }
-};
 
 // Fetch bookings from mock API
 const fetchBookings = async () => {
@@ -144,7 +115,6 @@ const cancelBooking = async (bookingId) => {
 };
 
 onMounted(() => {
-  fetchEvents();
   fetchBookings();
 });
 </script>
