+++ src/components/EventList.vue
@@ -1,0 +1,45 @@
+<template>
+  <section class="grid grid-cols-2 gap-8">
+    <template v-if="!loading">
+      <EventCard
+        v-for="event in events"
+        :key="event.id"
+        :title="event.title"
+        :when="event.date"
+        :description="event.description"
+        @register="$emit('register', event)"
+      />
+    </template>
+
+    <template v-else>
+      <LoadingEventCard v-for="i in 4" :key="i" />
+    </template>
+  </section>
+</template>
+
+<script setup>
+import { ref, onMounted } from 'vue';
+
+defineEmits(['register']);
+
+const events = ref([]);
+const loading = ref(false);
+
+import EventCard from '@/components/EventCard.vue';
+import LoadingEventCard from '@/components/LoadingEventCard.vue';
+
+// Fetch events from mock API
+const fetchEvents = async () => {
+  loading.value = true;
+  try {
+    const response = await fetch('http://localhost:3001/events');
+    events.value = await response.json();
+  } finally {
+    loading.value = false;
+  }
+};
+
+onMounted(() => {
+  fetchEvents();
+});
+</script>
